RF_requirements.txt
facexlib==0.2.5
ffmpeg-python==0.2.0
omegaconf==2.0.6
pytorch-lightning==1.0.8
basicsr==1.3.3.4
test-tube==0.7.5
opencv-python==4.5.2.54